[android] Clean up code
