+++ app_manager/appmanager/src/main/java/app/editors/manager/ui/activities/main/PasscodeActivity.kt
@@ -3,12 +3,11 @@ package app.editors.manager.ui.activities.main
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import androidx.activity.addCallback
 import androidx.activity.compose.setContent
 import app.editors.manager.app.App
 import app.editors.manager.managers.tools.PreferenceTool
 import app.editors.manager.ui.activities.base.BaseAppActivity
-import app.editors.manager.ui.compose.activities.main.PasscodeActivity
+import app.editors.manager.ui.compose.passcode.PasscodeMainScreen
 import lib.compose.ui.theme.ManagerTheme
 import javax.inject.Inject
 
@@ -31,19 +30,16 @@ class PasscodeActivity : BaseAppActivity() {
     @Inject
     lateinit var preferencesTool: PreferenceTool
 
-    var isEnterPasscode = false
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         App.getApp().appComponent.inject(this)
 
-        isEnterPasscode = intent.getBooleanExtra(ENTER_PASSCODE_KEY, false)
-        onBackPressedDispatcher.addCallback { finish() }
         setContent {
             ManagerTheme {
-                PasscodeActivity(
+                PasscodeMainScreen(
                     preferenceTool = preferencesTool,
+                    isEnterPasscode = intent.getBooleanExtra(ENTER_PASSCODE_KEY, false),
                     data = intent.extras,
                     backPressed = ::finish
                 )
